ЛП Большой взрыва (правки по код ревью (1))
