+++ src/components/templates/board-edit-event-template.js
@@ -1,4 +1,4 @@
-export const boardEditEvenTemplate = () => {
+export const boardEditEventTemplate = () => {
   return `<form class="event  event--edit" action="#" method="post">
                     <header class="event__header">
                       <div class="event__type-wrapper">
