+++ src/components/templates/board-event-template.js
