+++ src/components/templates/filters-template.js
